add score format setting

+++ lib/core/data/model/anime_list_item_model.dart
@@ -14,7 +14,7 @@ class MediaListItemModel with _$MediaListItemModel {
   factory MediaListItemModel({
     @Default('') String id,
     MediaListStatus? status,
-    int? score,
+    double? score,
     int? updatedAt,
     int? progress,
     int? progressVolumes,
