+++ lib/core/network/model/media_list_dto.dart
@@ -12,7 +12,7 @@ class MediaListDto with _$MediaListDto {
   factory MediaListDto({
     @Default(-1) @JsonKey(name: 'id') int id,
     @Default(-1) @JsonKey(name: 'userId') int userId,
-    @JsonKey(name: 'score') int? score,
+    @JsonKey(name: 'score') double? score,
     @JsonKey(name: 'status') MediaListStatus? status,
     @JsonKey(name: 'progress') int? progress,
     @JsonKey(name: 'progressVolumes') int? progressVolumes,
